Add time when playing (with flaws)
When in ttt, there will be a time that is displayed on top of the
board using SIGALRM.
However, the cursor jumps to the line behind "X>" and whenever the alarm
is signaled, the letter input disappears from the display. The game can
still be played as normal, but the user would not know the previous
input.